Move mock generation to mock.go and clean up old code
